+++ backend/src/main/java/com/deliveranything/domain/order/event/OrderCompletedEvent.java
@@ -1,14 +1,13 @@
 package com.deliveranything.domain.order.event;
 
 import com.deliveranything.domain.order.entity.Order;
-import java.math.BigDecimal;
 
 public record OrderCompletedEvent(
     Long orderId,
     Long riderProfileId,
     Long sellerProfileId,
-    BigDecimal storePrice,
-    BigDecimal deliveryPrice
+    Long storePrice,
+    Long deliveryPrice
 ) {
 
   public static OrderCompletedEvent fromOrder(Order order, Long riderId, Long sellerId) {
